+++ packages/block-library/src/image/test/use-open-image-media-editor-modal.js
@@ -1,8 +1,4 @@
 /**
- * @jest-environment jsdom
- */
-
-/**
  * External dependencies
  */
 import { act, renderHook } from '@testing-library/react';